+++ src/pywink/__init__.py
@@ -3,8 +3,13 @@ Top level functions
 """
 # noqa
 from pywink.api import set_bearer_token, refresh_access_token, \
-    set_wink_credentials, set_user_agent, get_bulbs, get_eggtrays, \
-    get_garage_doors, get_locks, get_powerstrip_outlets, get_sensors, \
-    get_shades, get_sirens, get_switches, get_devices, is_token_set, \
-    get_subscription_key, get_keys, get_piggy_banks, \
-    get_smoke_and_co_detectors, get_thermostats, get_set_access_token
+    set_wink_credentials, set_user_agent, wink_api_fetch, \
+    get_set_access_token, is_token_set, get_devices, \
+    get_subscription_key
+
+from pywink.api import get_bulbs, get_garage_doors, get_locks, \
+    get_powerstrip_outlets, get_shades, get_sirens, \
+    get_switches, get_thermostats
+
+from pywink.api import get_eggtrays, get_sensors, \
+    get_keys, get_piggy_banks, get_smoke_and_co_detectors