+++ src/pywink/domain/__init__.py
@@ -1,6 +1,0 @@
-"""
-Helper classes and functions for python-wink entities.
-
-Any behaviour which is not directly related to storing device state or
-interfacing with the Wink API should live in here.
-"""