+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Change Log
+## 0.10.1
+- Set the correct objectprefix for Wink Smoke/CO detectors and Piggy banks
+- Remove all desired state code
+
 ## 0.10.0
 - Support for Thermostats